+++ src/DataTransformer/Output/CompanyOutputDataTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\DataTransformer;
+namespace App\DataTransformer\Output;
 
 use ApiPlatform\Core\DataTransformer\DataTransformerInterface;
 use App\Dto\CompanyDtoOut;
