+++ src/DataTransformer/Output/InvoiceOutputDataTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\DataTransformer;
+namespace App\DataTransformer\Output;
 
 use ApiPlatform\Core\DataTransformer\DataTransformerInterface;
 use App\Dto\InvoiceDto;
